pyproject.toml
[project]
name = "suechef"
version = "0.1.0"
description = "SueChef - Your AI sous chef for cooking up winning legal strategies 👨‍🍳⚖️"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.6",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "eyecite>=2.7.5",
    "fastmcp>=2.5.2",
    "graphiti-core>=0.11.6",
    "neo4j>=5.28.1",
    "psycopg2-binary>=2.9.10",
    "qdrant-client>=1.14.2",
    "sentry-sdk>=2.29.1",
    "sqlalchemy>=2.0.41",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "pytest-mock>=3.14.1",
]
